backend/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.3
dj-database-url==0.4.2
Django==1.11.2
django-extensions==1.7.9
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
Markdown==2.6.8
pdb==0.1
pkg-resources==0.0.0
psycopg2==2.7.1
python-gnupg==0.4.0
pytz==2017.2
PyYAML==3.12
requests==2.17.3
six==1.10.0
urllib3==1.21.1
whitenoise==3.3.0
